% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_heart.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks of waves in raw PULSE data}
\usage{
find_peaks(t, y)
}
\arguments{
\item{t}{time}

\item{y}{val}
}
\description{
\code{heartbeatr-package} Find peaks of waves in raw PULSE data
}
