% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{Create a heartbeat instance}
\usage{
heartbeat(host = "127.0.0.1", port = 6379)
}
\arguments{
\item{host}{Hostname}

\item{port}{Port number}
}
\description{
Create a heartbeat instance.  This can be used by running
\code{obj$start(key, timeout)} which will reset the TTL on
\code{key} every \code{timeout} seconds (don't set this too high).
If the R process dies, then the key will expire after \code{3 *
timeout} seconds (or set \code{expire}) and another application can
tell that this R instance has died.
}
\details{
Heavily inspired by the \code{doRedis} package.
}

