% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat_send_signal}
\alias{heartbeat_send_signal}
\title{Send a signal}
\usage{
heartbeat_send_signal(con, key, signal)
}
\arguments{
\item{con}{A hiredis object}

\item{key}{The heartbeat key}

\item{signal}{A signal to send (e.g. \code{tools::SIGINT} or
\code{tools::SIGKILL})}
}
\description{
Sends a signal to a hearbeat process that is using key \code{key}
}

