% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeatr-package.R
\docType{package}
\name{heartbeatr-package}
\alias{heartbeatr}
\alias{heartbeatr-package}
\title{heartbeatr: A Workflow to Process Data Collected with PULSE Systems (Electricblue.eu)}
\description{
Given one or multiple paths to files produced by a PULSE multi-channel system (electricblue.eu/pulse) from a single experiment: [1] check pulse files for inconsistencies and read/merge all data, [2] split across time windows, [3] interpolate and smooth to optimize the raw data, [4] compute the heart rate frequency for each channel/window. Heart rate frequency is calculated using the XXXX algorithm.
}
\author{
\strong{Maintainer}: Rui Seabra \email{ptruimar@gmail.com} (\href{https://orcid.org/0000-0002-0240-3992}{ORCID})

Authors:
\itemize{
  \item Fernando Lima \email{fplima@gmail.com} (\href{https://orcid.org/0000-0001-9575-9834}{ORCID})
}

}
\keyword{internal}
