% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_plot.R
\name{pulse_anim}
\alias{pulse_anim}
\title{Plot and animate PULSE data}
\usage{
pulse_anim(heart_rates, folder = NULL, fps = 10, ID = NULL)
}
\arguments{
\item{heart_rates}{The output of \code{\link[=PULSE]{PULSE()}} (or \code{\link[=pulse_heart]{pulse_heart()}}, \code{\link[=pulse_doublecheck]{pulse_doublecheck()}}, \code{\link[=pulse_normalize]{pulse_normalize()}} and \code{\link[=pulse_summarise]{pulse_summarise()}}).}

\item{folder}{A valid path to an existing folder where the file \code{anim.mp4} will be saved. If not provided, \code{anim.mp4} will be saved to the current work directory. In any case, the final path is always printed at the end.}

\item{fps}{An integer representing the number of frames per second in the animation; defaults to \code{10}; more details below.}

\item{ID}{String naming a single or several target channel ids (each must match exactly); defaults to \code{NULL}, which results in all IDs being plotted}
}
\value{
An \code{.mp4} file is saved and its path printed to the console.
}
\description{
\strong{EXPERIMENTAL} - A shortcut function based on \code{ggplot2} and \code{gganimate} to facilitate the quick inspection of the quality of the analysis (with the peaks detected signaled with red dots). Useful to visually inspect the performance of the algorithm.
}
\details{
This function is \strong{NOT meant} for high-level displaying of the data - it's simply a quick shortcut to facilitate data inspection.

This function is \strong{EXPERIMENTAL}, and the resulting animation may not be correctly formatted in every machine - please be patient and provide constructive feedback.

On a machine running MacOS, the animation movie can easily be inspected using \verb{QuickTime Player}.

When inspecting the movie, go through each frame and assess if red dots top all heartbeats and no more than that. Additional info is displayed in the graph's title (such as sd). Difficult datasets may result in true heartbeats being missed or non-heartbeats (noise) being erroneously detected (false positives and/or false negatives). Note that the wider the time window (controlled by the \code{window_width_secs} parameter in \code{\link[=pulse_split]{pulse_split()}}) and the higher the heartbeat rate, the less critical are a few false positives or negatives (over a 10 secs window, missing 1 peak in 10 results in hz to drop by 10\% (from 1 to 0.9), while over a 30 secs window, missing 1 peak in 30 results in a drop of 3.33\% (from 1 to 0.967), and missing 1 peak in 60 results in a drop of just 1.7\%. With the animated plot, it becomes much quicker and easy to inspect the overall performance of the analysis over a large dataset.
}
\section{More on setting the right value for \code{fps}}{

The default value of \code{10} generates a shorter video with 10 individual graphs displayed over 1 second. At this rate, the resulting animation file is smaller, the animation plays quicker and detailed inspection requires pausing the animation playback (but frame-by-frame inspection is still possible). This is ideal for large datasets.
Alternatively, lower values will result in each individual graph persisting in the animation for a longer period, making it easier to inspect without the need to hit pause and play so often. Naturaly, the file size will also be larger, and if the dataset is large then the animation may extend for an uncomfortably long period of time. A value of \code{1} will result in only one graph being shown over each second.
}

\examples{
## Begin prepare data ----
paths <- pulse_example("RAW_original_")
heart_rates <- PULSE(
  paths,
  discard_channels  = paste0("s", 5:10),
  window_width_secs = 30,
  window_shift_secs = 60,
  min_data_points   = 0.8,
  interpolation_freq = 40,
  bandwidth   = 0.2,
  with_progress = TRUE
  )
## End prepare data ----

\dontrun{

# data from only one channel
pulse_anim(heart_rates, ID = "limpet_1")

# all channels
pulse_anim(heart_rates)
# notice how quickly one can inspect the entire dataset
#  and determine that only channel "limpet_1" has recorded good data
}
}
\seealso{
\itemize{
\item use \code{\link[=pulse_plot]{pulse_plot()}} to plot processed PULSE data for a several channels
\item \code{\link[=pulse_heart]{pulse_heart()}} or \code{\link[=PULSE]{PULSE()}} generate the input for \code{pulse_anim}
}
}
