% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_heart.R
\name{pulse_check}
\alias{pulse_check}
\title{Add stats to help identify possible doubling of hear rate frequencies (\verb{STEP 5})}
\usage{
pulse_check(heart_rates)
}
\arguments{
\item{heart_rates}{The output of \code{\link[=pulse_heart]{pulse_heart()}}}
}
\value{
A tibble similar to the one used as input, now augmented with two new columns: \code{dbl_rat} and \code{dbl_mag}.
}
\description{
\itemize{
\item \verb{step 1} -- \code{\link[=pulse_read]{pulse_read()}}
\item \verb{step 2} -- \code{\link[=pulse_split]{pulse_split()}}
\item \verb{step 3} -- \code{\link[=pulse_optimize]{pulse_optimize()}}
\item \verb{step 4} -- \code{\link[=pulse_heart]{pulse_heart()}}
\item \strong{\verb{-->>} step 5 -- \code{\link[=pulse_check]{pulse_check()}} \verb{<<--}}
}

Identify data points where it is likely that the heart rate frequency computed corresponds to double the actual heart rate frequency due to the algorithm having identified two peaks per heart beat
}
\examples{
# Begin prepare data ----
pulse_data_sub <- pulse_data
pulse_data_sub$data <- pulse_data_sub$data[,1:5]
pulse_data_split <- pulse_split(
   pulse_data_sub,
   window_width_secs = 30,
   window_shift_secs = 60,
   min_data_points = 0.8,
   with_progress = TRUE)
pulse_data_split <- pulse_optimize(pulse_data_split)
heart_rates <- pulse_heart(pulse_data_split)
# End prepare data ----

# Identify possible heart rate doublings
pulse_check(heart_rates)
}
\seealso{
\itemize{
\item \code{\link[=pulse_heart]{pulse_heart()}} generates the tibble to be used as input
\item \code{\link[=PULSE]{PULSE()}} is a wrapper function that executes all the steps needed to process PULSE data at once, including the identification of possible heart rate doublings
}
}
