% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_heart.R
\name{pulse_choose_keep}
\alias{pulse_choose_keep}
\title{Choose the best heart beat frequency estimate from among two estimates derived from raw and smoothed data}
\usage{
pulse_choose_keep(heart_rates)
}
\arguments{
\item{heart_rates}{The output from \code{\link[=pulse_heart]{pulse_heart()}}}
}
\value{
A tibble with the same structure as the input, but now with only one estimate for each combination of \code{id} and \code{time} (the one that was deemed better).
}
\description{
When running \code{\link[=pulse_optimize]{pulse_optimize()}} with \code{raw_v_smoothed = TRUE}, two estimates are generated for each data point, and \code{pulse_choose_keep} is used to automatically select the best one
}
\seealso{
\itemize{
\item \code{\link[=pulse_read]{pulse_read()}}, \code{\link[=pulse_split]{pulse_split()}}, \code{\link[=pulse_optimize]{pulse_optimize()}}, \code{\link[=pulse_heart]{pulse_heart()}} and \code{\link[=pulse_doublecheck]{pulse_doublecheck()}} are the other functions needed for the complete PULSE processing workflow
\item \code{\link[=PULSE]{PULSE()}} is a wrapper function that executes all the steps needed to process PULSE data at once
}
}
