% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pulse_data}
\alias{pulse_data}
\title{PULSE multi-channel example data}
\format{
\subsection{}{

A list with two elements:
\itemize{
\item \verb{$data}, a tibble with 29,978 rows and 11 columns - one column with timestamps (named \code{time}) and several columns of numeric data (the voltage readings from each channel of the PULSE multi-channel system; all channels with unique names)
\item \verb{$freq}, a single integer value representing the sampling frequency used (in Hz)
}
}
}
\usage{
pulse_data
}
\description{
A subset of data from an experiment monitoring the heartbeat of four \emph{Patella vulgata} limpets (channels S5-S10 were not monitored)
}
\keyword{datasets}
