% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pulse_example}
\alias{pulse_example}
\title{Get paths to pulse example files}
\usage{
pulse_example(pattern = NULL)
}
\arguments{
\item{pattern}{Pattern to select one or more example files. Pattern is vectorized, so more than one value can be supplied. If NULL, all example files are listed.}
}
\value{
The full path to one or more example files, or the filenames of all example files available.
}
\description{
\code{heartbeatr-package} comes bundled with several sample files in its inst/extdata directory. This function make them easy to access
}
\examples{
# Get the filenames of all example files
pulse_example()

# Get the full path to a target example file
pulse_example("RAW_original_20221229_1350.CSV")

# Get the full paths to any example files matching a search string
pulse_example("RAW_or")

# 'pulse_example()' is vectorized, meaning that multiple search strings can be used
pulse_example(c("RAW_or", "not_a_"))
}
\seealso{
\itemize{
\item \code{\link[=pulse_read]{pulse_read()}} can be used to read data from the example files
\item \code{\link[=PULSE]{PULSE()}} is a wrapper function that executes all the steps needed to process PULSE data at once, and it can be called to read and process the example files
}
}
