% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_heart.R
\name{pulse_find_peaks_one_channel}
\alias{pulse_find_peaks_one_channel}
\title{Determine the heart beat frequency in one PULSE channel}
\usage{
pulse_find_peaks_one_channel(split_window_one_channel)
}
\arguments{
\item{split_window_one_channel}{A tibble with PULSE data for only one channel with columns \verb{$time} and \verb{$val}}
}
\value{
A one-row tibble with 8 columns:
\itemize{
\item \code{time},  time at the center of split_window_one_channel$time
\item \code{t_pks}, time stamps of each wave peak identified
\item \code{n},     number of wave peaks identified
\item \code{sd},    standard deviation of the intervals between wave peaks
\item \code{hz},    heartbeat rate estimate (in Hz)
\item \code{ci},    confidence interval (hz ± ci)
}
}
\description{
Take data from one PULSE channel and identify the heartbeat wave peaks using an algorithm that searches for maxima across multiple scales.
}
\details{
function builds upon code from https://github.com/ig248/pyampd
}
\section{BPM}{

To convert to Beats Per Minute, simply multiply \code{hz} and \code{ci} by 60.
}

\examples{
## Begin prepare data ----
pulse_data_sub <- pulse_data
pulse_data_sub$data <- pulse_data_sub$data[,1:5]
pulse_data_split <- pulse_split(
   pulse_data_sub,
   window_width_secs = 30,
   window_shift_secs = 60,
   min_data_points = 0.8,
   with_progress = TRUE)
pulse_data_split <- pulse_optimize(pulse_data_split)
split_window <- pulse_data_split$data[[1]]
split_window_one_channel <- split_window[,1:2]
colnames(split_window_one_channel) <- c("time", "val")
# End prepare data ----

## Determine heartbeat rates in one channel in one time window
pulse_find_peaks_one_channel(split_window_one_channel)
}
\seealso{
\itemize{
\item \code{\link[=pulse_find_peaks_all_channels]{pulse_find_peaks_all_channels()}} runs \code{pulse_find_peaks_one_channel} on all PULSE channels
\item \code{\link[=pulse_read]{pulse_read()}}, \code{\link[=pulse_split]{pulse_split()}}, \code{\link[=pulse_optimize]{pulse_optimize()}}, \code{\link[=pulse_heart]{pulse_heart()}}, \code{\link[=pulse_doublecheck]{pulse_doublecheck()}} are the functions needed for the complete PULSE processing workflow
\item \code{\link[=PULSE]{PULSE()}} is a wrapper function that executes all the steps needed to process PULSE data at once
}
}
