% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_plot.R
\name{pulse_plot_all}
\alias{pulse_plot_all}
\title{Plot processed PULSE data}
\usage{
pulse_plot_all(
  heart_rates,
  normalized = FALSE,
  smooth = TRUE,
  points = TRUE,
  facets = TRUE
)
}
\arguments{
\item{heart_rates}{The output of \code{\link[=PULSE]{PULSE()}} (or \code{\link[=pulse_heart]{pulse_heart()}}, \code{\link[=pulse_normalize]{pulse_normalize()}} and \code{\link[=pulse_summarise]{pulse_summarise()}}).}

\item{normalized}{Logical, whether to plot \code{hz_norm} (\code{TRUE}) or \code{hz} (\code{FALSE}, the default).}

\item{smooth}{Logical, whether to plot a \code{loess} smoothing curve (\code{TRUE}, the default) or a line (\code{FALSE}).}

\item{points}{Logical, whether to plot the data points (defaults to \code{TRUE}).}

\item{facets}{Logical, whether to separate channels in facets (\code{TRUE}, the default) or to plot all data together (\code{FALSE})}
}
\value{
A ggplot object that can be augmented using ggplot2 syntax or plotted right away
}
\description{
A shortcut function based on \code{ggplot2} to facilitate the quick inspection of the results of the analysis (with the options to separate channels or not).
}
\details{
This function is \strong{NOT meant} for high-level displaying of the data - it's simply a quick shortcut to facilitate data inspection.

When inspecting the plot, assess if the \code{loess} confidence intervals are too wide for any given channel, which is indicative of data with high variability (usually not desired).
}
\examples{
# Begin prepare data ----
paths <- pulse_example("RAW_original_")
heart_rates <- PULSE(
  paths,
  discard_channels  = paste0("s", 5:10),
  window_width_secs = 30,
  window_shift_secs = 60,
  min_data_points   = 0.8,
  target_freq = 40,
  bandwidth   = 0.2,
  with_progress = TRUE
  )
# End prepare data ----

# Default
pulse_plot_all(heart_rates)

# Without facets, the different basal heartbeat rates become evident in non-normalized data
pulse_plot_all(heart_rates, facets = FALSE)

# Without facets, normalized data always lines up (around 1) during the baseline period
pulse_plot_all(pulse_normalize(heart_rates), normalize = TRUE, facets = FALSE)

# The plot can be modified using ggplot2 syntax
pulse_plot_all(heart_rates) + ggplot2::theme_classic()
}
\seealso{
\code{\link[=pulse_plot_raw]{pulse_plot_raw()}}, \code{\link[=PULSE]{PULSE()}}, \code{\link[=pulse_heart]{pulse_heart()}}, \code{\link[=pulse_normalize]{pulse_normalize()}}
}
