% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_read.R
\name{pulse_read_checks}
\alias{pulse_read_checks}
\title{before pulse_read}
\usage{
pulse_read_checks(paths)
}
\arguments{
\item{paths}{File paths to CSV files produced by a PULSE system during a single experiment.}
}
\value{
_
}
\description{
before pulse_read
}
\examples{
_
}
