% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_read.R
\name{pulse_read_data}
\alias{pulse_read_data}
\title{heartbeatr utility function}
\usage{
pulse_read_data(path, skip, cols, multi)
}
\arguments{
\item{path}{file path}

\item{skip}{numeric}

\item{cols}{colnames}

\item{multi}{logical}
}
\description{
\code{heartbeatr-package} utility function
}
