% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{Create a heartbeat instance}
\usage{
heartbeat(key, period, expire = 3 * period, value = expire,
  host = "localhost", port = 6379L, password = NULL, db = NULL,
  start = TRUE)
}
\arguments{
\item{key}{Key to use}

\item{period}{Timeout period (in seconds)}

\item{expire}{Key expiry time (in seconds)}

\item{value}{Value to store in the key.  By default it stores the
expiry time, so the time since last heartbeat can be computed.}

\item{host}{Redis host to use (by default localhost)}

\item{port}{Redis port to use (by default 6379)}

\item{password}{Optional password used (via the \code{AUTH} command
before any redis commands are run on the server}

\item{db}{Database to connect to (if not the default).}

\item{start}{Should the heartbeat be started immediately?}
}
\description{
Create a heartbeat instance.  This can be used by running
\code{obj$start()} which will reset the TTL on \code{key} every
\code{period} seconds (don't set this too high).  If the R process
dies, then the key will expire after \code{3 * period} seconds (or
set \code{expire}) and another application can tell that this R
instance has died.
}
\details{
The heartbeat object has three methods:
\itemize{

\item \code{is_running()} which returns \code{TRUE} or
\code{FALSE} if the heartbeat is/is not running.

\item \code{start()} which starts a heartbeat

\item \code{stop()} which requess a stops for the heartbeat

}

Heavily inspired by the \code{doRedis} package.
}

