% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat_}
\alias{heartbeat_}
\alias{heartbeat}
\title{Create a heartbeat instance}
\usage{
heartbeat(
  key,
  period,
  expire = 3 * period,
  value = expire,
  config = NULL,
  start = TRUE,
  timeout = 10
)
}
\arguments{
\item{key}{Key to use}

\item{period}{Timeout period (in seconds)}

\item{expire}{Key expiry time (in seconds)}

\item{value}{Value to store in the key.  By default it stores the
expiry time, so the time since last heartbeat can be computed.}

\item{config}{Configuration parameters passed through to
\code{redux::redis_config}.  Provide as either a named list or a
\code{redis_config} object.  This allows host, port, password,
db, etc all to be set.  Socket connections (i.e., using
\code{path} to access Redis over a socket) are not currently
supported.}

\item{start}{Should the heartbeat be started immediately?}

\item{timeout}{Time, in seconds, to wait for the heartbeat to
appear.  It should generally appear very quickly (within a
second unless your connection is very slow) so this can be
generally left alone.}
}
\description{
Create a heartbeat instance.  This can be used by running
\code{obj$start()} which will reset the TTL (Time To Live) on \code{key} every
\code{period} seconds (don't set this too high).  If the R process
dies, then the key will expire after \code{3 * period} seconds (or
set \code{expire}) and another application can tell that this R
instance has died.
}
\examples{

if (redux::redis_available()) {
  rand_str <- function() {
    paste(sample(letters, 20, TRUE), collapse = "")
  }
  key <- sprintf("heartbeatr:test:\%s", rand_str())
  h <- heartbeatr::heartbeat(key, 1, expire = 2)
  con <- redux::hiredis()

  # The heartbeat key exists
  con$EXISTS(key)

  # And has an expiry of less than 2000ms
  con$PTTL(key)

  # We can manually stop the heartbeat, and 2s later the key will
  # stop existing
  h$stop()

  Sys.sleep(2)
  con$EXISTS(key) # 0

  # This is required to close any processes opened by this
  # example, normally you would not need this.
  processx:::supervisor_kill()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{heartbeat_$new()}}
\item \href{#method-is_running}{\code{heartbeat_$is_running()}}
\item \href{#method-start}{\code{heartbeat_$start()}}
\item \href{#method-stop}{\code{heartbeat_$stop()}}
\item \href{#method-format}{\code{heartbeat_$format()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a heartbeat object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{heartbeat_$new(config, key, value, period, expire, start = FALSE, timeout = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Configuration parameters passed through to
\code{redux::redis_config}.  Provide as either a named list or a
\code{redis_config} object.  This allows host, port, password,
db, etc all to be set.  Socket connections (i.e., using
\code{path} to access Redis over a socket) are not currently
supported.}

\item{\code{key}}{Key to use}

\item{\code{value}}{Value to store in the key.  By default it stores the
expiry time, so the time since last heartbeat can be computed.}

\item{\code{period}}{Timeout period (in seconds)}

\item{\code{expire}}{Key expiry time (in seconds)}

\item{\code{start}}{Should the heartbeat be started immediately?}

\item{\code{timeout}}{Time, in seconds, to wait for the heartbeat to
appear.  It should generally appear very quickly (within a
second unless your connection is very slow) so this can be
generally left alone.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-is_running}{}}}
\subsection{Method \code{is_running()}}{
Report if heartbeat process is running. This will be
\code{TRUE} if the process has been started and has not stopped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{heartbeat_$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
Start the heartbeat process. An error will be thrown
if it is already running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{heartbeat_$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop the heartbeat process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{heartbeat_$stop(wait = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wait}}{Logical, indicating if we should wait until the
heartbeat process terminates (should take only a fraction of a
second)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
Format method, used by R6 to nicely print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{heartbeat_$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments, currently ignored}
}
\if{html}{\out{</div>}}
}
}
}
